+++ packages/camera/CHANGELOG.md
@@ -13,5 +13,6 @@
 ## 0.3.0
 
 * Update camera to 0.9.4.
+* Implement `pausePreview` and `resumePreview`.
 * Update the example app and integration_test.
 * Remove the unused test driver.